Add backend dev server tooling
