+++ .bazelignore
@@ -23,3 +23,4 @@ target
 jest.config.js
 eslint.config.mjs
 libs/ballot-interpreter/src/hmpb-ts/rust_addon.node
+libs/pdi-scanner/src/ts/pdictl
