+++ apps/design/backend/BUILD.bazel
@@ -1,4 +1,13 @@
 load("@aspect_rules_js//js:defs.bzl", "js_library")
+load("//tools/apps:dev_backend.bzl", "dev_backend")
+
+dev_backend(
+    name = "dev_server",
+    backend_entry_point_from_root = package_name() + "/src/main.ts",
+    backend_path_from_root = package_name(),
+    port = "3002",
+    vx_machine_type = "design",
+)
 
 js_library(
     name = "db_schema",
