+++ libs/ballot-interpreter/src/BUILD.bazel
@@ -7,7 +7,7 @@ load("//tools/ts_build:ts_tests.bzl", "ts_tests")
 
 ts_library(
     name = "src",
-    data = [":rust_addon_copy"],
+    data = [":rust_addon"],
     deps = [
         "//:node_modules/@types/better-sqlite3",
         "//:node_modules/@types/chalk",
@@ -64,9 +64,13 @@ ts_tests(
 )
 
 copy_file(
-    name = "rust_addon_copy",
+    name = "rust_addon",
     src = "//libs/ballot-interpreter:rust",
     out = "hmpb-ts/rust_addon.node",
+    visibility = [
+        ":__pkg__",
+        "//apps/mark-scan/backend:__pkg__",  # Used when running the dev server.
+    ],
 )
 
 lint_test(name = "lint")
