+++ libs/pdi-scanner/src/ts/pdictl
@@ -1,0 +1,1 @@
+../../../../bazel-bin/libs/pdi-scanner/src/ts/pdictl
