+++ tools/tsx/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@npm//:tsx/package_json.bzl", tsx = "bin")
+
+tsx.tsx_binary(
+    name = "cli",
+    env = {
+        "BAZEL_BINDIR": ".",
+    },
+    tags = ["manual"],
+    visibility = ["//visibility:public"],
+)
